+++ app/core/Engine/messengers/currency-rate-controller-messenger/currency-rate-controller-messenger.ts
@@ -1,4 +1,4 @@
-import { CurrencyRateMessenger } from '../../controllers/CurrencyRateController';
+import { CurrencyRateMessenger } from '../../controllers/currency-rate-controller/currency-rate-controller-init';
 import { BaseControllerMessenger } from '../../types';
 
 /**
