+++ app/core/Engine/messengers/multichain-assets-rates-controller-messenger/multichain-assets-rates-controller-messenger.ts
@@ -1,9 +1,6 @@
 import { BaseControllerMessenger } from '../../types';
 import { MultichainAssetsRatesControllerMessenger } from '@metamask/assets-controllers';
 
-// Export the types
-export * from './types';
-
 /**
  * Get the MultichainAssetsRatesControllerMessenger for the MultichainAssetsRatesController.
  *
