+++ app/core/Engine/messengers/multichain-balances-controller-messenger/multichain-balances-controller-messenger.ts
@@ -1,9 +1,6 @@
 import { BaseControllerMessenger } from '../../types';
 import { MultichainBalancesControllerMessenger } from '@metamask/assets-controllers';
 
-// Export the types
-export * from './types';
-
 /**
  * Get the MultichainBalancesControllerMessenger for the MultichainBalancesController.
  *
